@@ -43,7 +43,6 @@ from pytensor.tensor.extra_ops import CumOp
 
 from pymc.logprob.abstract import MeasurableVariable, _logprob, _logprob_helper
 from pymc.logprob.rewriting import PreserveRVMappings, measurable_ir_rewrites_db
-from pymc.logprob.utils import ignore_logprob
 
 
 class MeasurableCumsum(CumOp):
@@ -92,25 +91,17 @@ def find_measurable_cumsums(fgraph, node) -> Optional[List[MeasurableCumsum]]:
     if rv_map_feature is None:
         return None  # pragma: no cover
 
-    rv = node.outputs[0]
-
     base_rv = node.inputs[0]
-    if not (
-        base_rv.owner
-        and isinstance(base_rv.owner.op, MeasurableVariable)
-        and base_rv not in rv_map_feature.rv_values
-    ):
-        return None  # pragma: no cover
 
     # Check that cumsum does not mix dimensions
     if base_rv.ndim > 1 and node.op.axis is None:
         return None
 
+    if not rv_map_feature.request_measurable(node.inputs):
+        return None
+
     new_op = MeasurableCumsum(axis=node.op.axis or 0, mode="add")
-    # Make base_var unmeasurable
-    unmeasurable_base_rv = ignore_logprob(base_rv)
-    new_rv = new_op.make_node(unmeasurable_base_rv).default_output()
-    new_rv.name = rv.name
+    new_rv = new_op.make_node(base_rv).default_output()
 
     return [new_rv]
 
