@@ -42,15 +42,11 @@ import pytensor.tensor as pt
 from pytensor.graph.basic import Apply, Constant, Variable
 from pytensor.graph.fg import FunctionGraph
 from pytensor.graph.op import Op, compute_test_value
-from pytensor.graph.rewriting.basic import (
-    EquilibriumGraphRewriter,
-    node_rewriter,
-    pre_greedy_node_rewriter,
-)
+from pytensor.graph.rewriting.basic import node_rewriter, pre_greedy_node_rewriter
 from pytensor.ifelse import IfElse, ifelse
-from pytensor.scalar.basic import Switch
-from pytensor.tensor.basic import Join, MakeVector
-from pytensor.tensor.elemwise import Elemwise
+from pytensor.scalar import Switch
+from pytensor.scalar import switch as scalar_switch
+from pytensor.tensor.basic import Join, MakeVector, switch
 from pytensor.tensor.random.rewriting import (
     local_dimshuffle_rv_lift,
     local_rv_size_lift,
@@ -61,7 +57,6 @@ from pytensor.tensor.subtensor import (
     AdvancedSubtensor,
     AdvancedSubtensor1,
     as_index_literal,
-    as_nontensor_scalar,
     get_canonical_form_slice,
     is_basic_idx,
 )
@@ -69,15 +64,21 @@ from pytensor.tensor.type import TensorType
 from pytensor.tensor.type_other import NoneConst, NoneTypeT, SliceConstant, SliceType
 from pytensor.tensor.var import TensorVariable
 
-from pymc.logprob.abstract import MeasurableVariable, _logprob, _logprob_helper
+from pymc.logprob.abstract import (
+    MeasurableElemwise,
+    MeasurableVariable,
+    _logprob,
+    _logprob_helper,
+)
 from pymc.logprob.rewriting import (
+    PreserveRVMappings,
+    assume_measured_ir_outputs,
     local_lift_DiracDelta,
-    logprob_rewrites_db,
     measurable_ir_rewrites_db,
     subtensor_ops,
 )
 from pymc.logprob.tensor import naive_bcast_rv_lift
-from pymc.logprob.utils import ignore_logprob, ignore_logprob_multiple_vars
+from pymc.logprob.utils import check_potential_measurability
 
 
 def is_newaxis(x):
@@ -272,7 +273,7 @@ def get_stack_mixture_vars(
 
 
 @node_rewriter(subtensor_ops)
-def mixture_replace(fgraph, node):
+def find_measurable_index_mixture(fgraph, node):
     r"""Identify mixture sub-graphs and replace them with a place-holder `Op`.
 
     The basic idea is to find ``stack(mixture_comps)[I_rv]``, where
@@ -281,28 +282,9 @@ def mixture_replace(fgraph, node):
     From these terms, new terms ``Z_rv[i] = mixture_comps[i][i == I_rv]`` are
     created for each ``i`` in ``enumerate(mixture_comps)``.
     """
-    rv_map_feature = getattr(fgraph, "preserve_rv_mappings", None)
+    rv_map_feature: Optional[PreserveRVMappings] = getattr(fgraph, "preserve_rv_mappings", None)
 
     if rv_map_feature is None:
-        return None  # pragma: no cover
-
-    old_mixture_rv = node.default_output()
-
-    mixture_rvs, join_axis = get_stack_mixture_vars(node)
-
-    # We don't support symbolic join axis
-    if mixture_rvs is None or not isinstance(join_axis, (NoneTypeT, Constant)):
-        return None
-
-    # Check that all components are MeasurableVariables and none is already conditioned on
-    if not all(
-        (
-            rv.owner is not None
-            and isinstance(rv.owner.op, MeasurableVariable)
-            and rv not in rv_map_feature.rv_values
-        )
-        for rv in mixture_rvs
-    ):
         return None  # pragma: no cover
 
     mixing_indices = node.inputs[1:]
@@ -318,15 +300,15 @@ def mixture_replace(fgraph, node):
         ):
             return None
 
-    # We loop through mixture components and collect all the array elements
-    # that belong to each one (by way of their indices).
-    new_mixture_rvs = []
-    for i, component_rv in enumerate(mixture_rvs):
-        # We create custom types for the mixture components and assign them
-        # null `get_measurable_outputs` dispatches so that they aren't
-        # erroneously encountered in places like `factorized_joint_logprob`.
-        new_comp_rv = ignore_logprob(component_rv)
-        new_mixture_rvs.append(new_comp_rv)
+    old_mixture_rv = node.default_output()
+    mixture_rvs, join_axis = get_stack_mixture_vars(node)
+
+    # We don't support symbolic join axis
+    if mixture_rvs is None or not isinstance(join_axis, (NoneTypeT, Constant)):
+        return None
+
+    if rv_map_feature.request_measurable(mixture_rvs) != mixture_rvs:
+        return None
 
     # Replace this sub-graph with a `MixtureRV`
     mix_op = MixtureRV(
@@ -334,7 +316,7 @@ def mixture_replace(fgraph, node):
         old_mixture_rv.dtype,
         old_mixture_rv.broadcastable,
     )
-    new_node = mix_op.make_node(*([join_axis] + mixing_indices + new_mixture_rvs))
+    new_node = mix_op.make_node(*([join_axis] + mixing_indices + mixture_rvs))
 
     new_mixture_rv = new_node.default_output()
 
@@ -345,55 +327,6 @@ def mixture_replace(fgraph, node):
             compute_test_value(node)
 
         new_mixture_rv.tag.test_value = old_mixture_rv.tag.test_value
-
-    if old_mixture_rv.name:
-        new_mixture_rv.name = f"{old_mixture_rv.name}-mixture"
-
-    return [new_mixture_rv]
-
-
-@node_rewriter((Elemwise,))
-def switch_mixture_replace(fgraph, node):
-    rv_map_feature = getattr(fgraph, "preserve_rv_mappings", None)
-
-    if rv_map_feature is None:
-        return None  # pragma: no cover
-
-    if not isinstance(node.op.scalar_op, Switch):
-        return None  # pragma: no cover
-
-    old_mixture_rv = node.default_output()
-    # idx, component_1, component_2 = node.inputs
-
-    mixture_rvs = []
-
-    for component_rv in node.inputs[1:]:
-        if not (
-            component_rv.owner
-            and isinstance(component_rv.owner.op, MeasurableVariable)
-            and component_rv not in rv_map_feature.rv_values
-        ):
-            return None
-        new_comp_rv = ignore_logprob(component_rv)
-        mixture_rvs.append(new_comp_rv)
-
-    mix_op = MixtureRV(
-        2,
-        old_mixture_rv.dtype,
-        old_mixture_rv.broadcastable,
-    )
-    new_node = mix_op.make_node(*([NoneConst, as_nontensor_scalar(node.inputs[0])] + mixture_rvs))
-
-    new_mixture_rv = new_node.default_output()
-
-    if pytensor.config.compute_test_value != "off":
-        if not hasattr(old_mixture_rv.tag, "test_value"):
-            compute_test_value(node)
-
-        new_mixture_rv.tag.test_value = old_mixture_rv.tag.test_value
-
-    if old_mixture_rv.name:
-        new_mixture_rv.name = f"{old_mixture_rv.name}-mixture"
 
     return [new_mixture_rv]
 
@@ -475,12 +408,61 @@ def logprob_MixtureRV(
     return logp_val
 
 
-logprob_rewrites_db.register(
-    "mixture_replace",
-    EquilibriumGraphRewriter(
-        [mixture_replace, switch_mixture_replace],
-        max_use_ratio=pytensor.config.optdb__max_use_ratio,
-    ),
+class MeasurableSwitchMixture(MeasurableElemwise):
+    valid_scalar_types = (Switch,)
+
+
+measurable_switch_mixture = MeasurableSwitchMixture(scalar_switch)
+
+
+@node_rewriter([switch])
+def find_measurable_switch_mixture(fgraph, node):
+    rv_map_feature: Optional[PreserveRVMappings] = getattr(fgraph, "preserve_rv_mappings", None)
+
+    if rv_map_feature is None:
+        return None  # pragma: no cover
+
+    switch_cond, *components = node.inputs
+
+    # We don't support broadcasting of components, as that yields dependent (identical) values.
+    # The current logp implementation assumes all component values are independent.
+    # Broadcasting of the switch condition is fine
+    out_bcast = node.outputs[0].type.broadcastable
+    if any(comp.type.broadcastable != out_bcast for comp in components):
+        return None
+
+    # Check that `switch_cond` is not potentially measurable
+    valued_rvs = rv_map_feature.rv_values.keys()
+    if check_potential_measurability([switch_cond], valued_rvs):
+        return None
+
+    if rv_map_feature.request_measurable(components) != components:
+        return None
+
+    return [measurable_switch_mixture(switch_cond, *components)]
+
+
+@_logprob.register(MeasurableSwitchMixture)
+def logprob_switch_mixture(op, values, switch_cond, component_true, component_false, **kwargs):
+    [value] = values
+
+    return switch(
+        switch_cond,
+        _logprob_helper(component_true, value),
+        _logprob_helper(component_false, value),
+    )
+
+
+measurable_ir_rewrites_db.register(
+    "find_measurable_index_mixture",
+    find_measurable_index_mixture,
+    "basic",
+    "mixture",
+)
+
+measurable_ir_rewrites_db.register(
+    "find_measurable_switch_mixture",
+    find_measurable_switch_mixture,
     "basic",
     "mixture",
 )
@@ -494,31 +476,65 @@ MeasurableVariable.register(MeasurableIfElse)
 
 
 @node_rewriter([IfElse])
+def useless_ifelse_outputs(fgraph, node):
+    """Remove outputs that are shared across the IfElse branches."""
+    # TODO: This should be a PyTensor canonicalization
+    op = node.op
+    if_var, *inputs = node.inputs
+    shared_inputs = set(inputs[op.n_outs :]).intersection(inputs[: op.n_outs])
+    if not shared_inputs:
+        return None
+
+    replacements = {}
+    for shared_inp in shared_inputs:
+        idx = inputs.index(shared_inp)
+        replacements[node.outputs[idx]] = shared_inp
+
+    # IfElse isn't needed at all
+    if len(shared_inputs) == op.n_outs:
+        return replacements
+
+    # Create subset IfElse with remaining nodes
+    remaining_inputs = [inp for inp in inputs if inp not in shared_inputs]
+    new_outs = (
+        IfElse(n_outs=len(remaining_inputs) // 2).make_node(if_var, *remaining_inputs).outputs
+    )
+    for inp, new_out in zip(remaining_inputs, new_outs):
+        idx = inputs.index(inp)
+        replacements[node.outputs[idx]] = new_out
+
+    return replacements
+
+
+@node_rewriter([IfElse])
 def find_measurable_ifelse_mixture(fgraph, node):
-    rv_map_feature = getattr(fgraph, "preserve_rv_mappings", None)
+    rv_map_feature: Optional[PreserveRVMappings] = getattr(fgraph, "preserve_rv_mappings", None)
 
     if rv_map_feature is None:
         return None  # pragma: no cover
 
-    if isinstance(node.op, MeasurableIfElse):
-        return None
-
-    # Check if all components are unvalued measuarable variables
+    op = node.op
     if_var, *base_rvs = node.inputs
 
-    if not all(
-        (
-            rv.owner is not None
-            and isinstance(rv.owner.op, MeasurableVariable)
-            and rv not in rv_map_feature.rv_values
-        )
-        for rv in base_rvs
-    ):
-        return None  # pragma: no cover
+    valued_rvs = rv_map_feature.rv_values.keys()
+    if not all(check_potential_measurability([base_var], valued_rvs) for base_var in base_rvs):
+        return None
 
-    unmeasurable_base_rvs = ignore_logprob_multiple_vars(base_rvs, rv_map_feature.rv_values)
+    base_rvs = assume_measured_ir_outputs(valued_rvs, base_rvs)
+    if len(base_rvs) != op.n_outs * 2:
+        return None
+    if not all(var.owner and isinstance(var.owner.op, MeasurableVariable) for var in base_rvs):
+        return None
 
-    return MeasurableIfElse(n_outs=node.op.n_outs).make_node(if_var, *unmeasurable_base_rvs).outputs
+    return MeasurableIfElse(n_outs=op.n_outs).make_node(if_var, *base_rvs).outputs
+
+
+measurable_ir_rewrites_db.register(
+    "useless_ifelse_outputs",
+    useless_ifelse_outputs,
+    "basic",
+    "mixture",
+)
 
 
 measurable_ir_rewrites_db.register(
@@ -551,4 +567,7 @@ def logprob_ifelse(op, values, if_var, *base_rvs, **kwargs):
     logps_then = replace_rvs_by_values(logps_then, rvs_to_values=rvs_to_values_then)
     logps_else = replace_rvs_by_values(logps_else, rvs_to_values=rvs_to_values_else)
 
-    return ifelse(if_var, logps_then, logps_else)
+    logps = ifelse(if_var, logps_then, logps_else)
+    if len(logps) == 1:
+        return logps[0]
+    return logps
