@@ -115,7 +115,7 @@ def get_jaxified_graph(
     # JAX sequential optimizer without warnings. We made sure there
     # are no mutable input variables, so we only need to check for
     # "destroyers". This should be automatically handled by PyTensor
-    # once https://github.com/pytensor-devs/pytensor/issues/637 is fixed.
+    # once https://github.com/aesara-devs/aesara/issues/637 is fixed.
     fgraph.attach_feature(
         Supervisor(
             input
@@ -317,7 +317,6 @@ def sample_blackjax_nuts(
     postprocessing_backend: Optional[str] = None,
     postprocessing_chunks: Optional[int] = None,
     idata_kwargs: Optional[Dict[str, Any]] = None,
-    **kwargs,
 ) -> az.InferenceData:
     """
     Draw samples from the posterior using the NUTS method from the ``blackjax`` library.
@@ -530,7 +529,6 @@ def sample_numpyro_nuts(
     postprocessing_chunks: Optional[int] = None,
     idata_kwargs: Optional[Dict] = None,
     nuts_kwargs: Optional[Dict] = None,
-    **kwargs,
 ) -> az.InferenceData:
     """
     Draw samples from the posterior using the NUTS method from the ``numpyro`` library.
