@@ -105,9 +105,11 @@ def from_yaml():
     _log.info("Number of test runs (❌=0, ✅=once)\n%s", df.replace(0, "❌").replace(1, "✅"))
 
     if ignored_by_all:
-        _log.warning("%i tests are completely ignored:\n%s", len(ignored_by_all), ignored_by_all)
+        raise AssertionError(
+            f"{len(ignored_by_all)} tests are completely ignored:\n{ignored_by_all}"
+        )
     if run_multiple_times:
-        raise Exception(
+        raise AssertionError(
             f"{len(run_multiple_times)} tests are run multiple times with the same OS and floatX setting:\n{run_multiple_times}"
         )
     return
