@@ -40,9 +40,9 @@ import pytensor.tensor as pt
 import pytest
 import scipy.stats as st
 
-from pymc import logp
+from pymc import draw, logp
 from pymc.logprob.abstract import MeasurableVariable
-from pymc.logprob.basic import factorized_joint_logprob
+from pymc.logprob.basic import conditional_logp
 from pymc.logprob.censoring import MeasurableClip
 from pymc.logprob.rewriting import construct_ir_fgraph
 from pymc.testing import assert_no_rvs
@@ -64,7 +64,7 @@ def test_scalar_clipped_mixture():
     idxs_vv = idxs.clone()
     idxs_vv.name = "idxs_val"
 
-    logp = factorized_joint_logprob({idxs: idxs_vv, mix: mix_vv})
+    logp = conditional_logp({idxs: idxs_vv, mix: mix_vv})
     logp_combined = pt.sum([pt.sum(factor) for factor in logp.values()])
 
     logp_fn = pytensor.function([idxs_vv, mix_vv], logp_combined)
@@ -102,9 +102,7 @@ def test_nested_scalar_mixtures():
     idxs12_vv = idxs12.clone()
     mix12_vv = mix12.clone()
 
-    logp = factorized_joint_logprob(
-        {idxs1: idxs1_vv, idxs2: idxs2_vv, idxs12: idxs12_vv, mix12: mix12_vv}
-    )
+    logp = conditional_logp({idxs1: idxs1_vv, idxs2: idxs2_vv, idxs12: idxs12_vv, mix12: mix12_vv})
     logp_combined = pt.sum([pt.sum(factor) for factor in logp.values()])
 
     logp_fn = pytensor.function([idxs1_vv, idxs2_vv, idxs12_vv, mix12_vv], logp_combined)
@@ -139,7 +137,7 @@ def test_unvalued_ir_reversion(nested):
 
     z_fgraph, _, memo = construct_ir_fgraph(rv_values)
 
-    assert len(z_fgraph.preserve_rv_mappings.measurable_conversions) == 1 + nested
+    # assert len(z_fgraph.preserve_rv_mappings.measurable_conversions) == 1
     assert (
         sum(isinstance(node.op, MeasurableVariable) for node in z_fgraph.apply_nodes) == 2
     )  # Just the 2 rvs
@@ -217,4 +215,50 @@ def test_affine_log_transform_rv():
     assert np.allclose(
         logp_fn(a_val, b_val, y_val),
         st.norm(a_val, b_val).logpdf(y_val),
+    )
+
+
+@pytest.mark.parametrize("reverse", (False, True))
+def test_affine_join_interdependent(reverse):
+    x = pt.random.normal(name="x")
+    y_rvs = []
+    prev_rv = x
+    for i in range(3):
+        next_rv = pt.exp(prev_rv + pt.random.beta(3, 1, name=f"y{i}", size=(1, 2)))
+        y_rvs.append(next_rv)
+        prev_rv = next_rv
+
+    if reverse:
+        y_rvs = y_rvs[::-1]
+
+    ys = pt.concatenate(y_rvs, axis=0)
+    ys.name = "ys"
+
+    x_vv = x.clone()
+    ys_vv = ys.clone()
+
+    logp = conditional_logp({x: x_vv, ys: ys_vv})
+    logp_combined = pt.sum([pt.sum(factor) for factor in logp.values()])
+    assert_no_rvs(logp_combined)
+
+    y0_vv = y_rvs[0].clone()
+    y1_vv = y_rvs[1].clone()
+    y2_vv = y_rvs[2].clone()
+
+    ref_logp = conditional_logp({x: x_vv, y_rvs[0]: y0_vv, y_rvs[1]: y1_vv, y_rvs[2]: y2_vv})
+    ref_logp_combined = pt.sum([pt.sum(factor) for factor in ref_logp.values()])
+
+    rng = np.random.default_rng()
+    x_vv_test, ys_vv_test = draw([x, ys], random_seed=rng)
+    ys_vv_test = rng.normal(size=(3, 2))
+    np.testing.assert_allclose(
+        logp_combined.eval({x_vv: x_vv_test, ys_vv: ys_vv_test}),
+        ref_logp_combined.eval(
+            {
+                x_vv: x_vv_test,
+                y0_vv: ys_vv_test[0:1],
+                y1_vv: ys_vv_test[1:2],
+                y2_vv: ys_vv_test[2:3],
+            }
+        ),
     )
