@@ -45,7 +45,7 @@ from pytensor.scan.utils import ScanArgs
 from scipy import stats
 
 from pymc.logprob.abstract import _logprob_helper
-from pymc.logprob.basic import factorized_joint_logprob, logp
+from pymc.logprob.basic import conditional_logp, logp
 from pymc.logprob.scan import (
     construct_scan,
     convert_outer_out_to_in,
@@ -335,7 +335,7 @@ def test_scan_joint_logprob(require_inner_rewrites):
     s_vv = S_rv.clone()
     s_vv.name = "s"
 
-    y_logp = factorized_joint_logprob({Y_rv: y_vv, S_rv: s_vv, Gamma_rv: Gamma_vv})
+    y_logp = conditional_logp({Y_rv: y_vv, S_rv: s_vv, Gamma_rv: Gamma_vv})
     y_logp_combined = pt.sum([pt.sum(factor) for factor in y_logp.values()])
 
     y_val = np.arange(10)
@@ -421,7 +421,7 @@ def test_initial_values():
     s_1T_vv = S_1T_rv.clone()
     s_1T_vv.name = "s_1T"
 
-    logp_parts = factorized_joint_logprob({S_1T_rv: s_1T_vv, S_0_rv: s_0_vv})
+    logp_parts = conditional_logp({S_1T_rv: s_1T_vv, S_0_rv: s_0_vv})
 
     s_0_val = 0
     s_1T_val = np.array([1, 0, 1, 0, 1, 1, 0, 1, 0, 1])
@@ -492,7 +492,7 @@ def test_scan_over_seqs():
 
     xs_vv = ys.clone()
     ys_vv = ys.clone()
-    ys_logp = factorized_joint_logprob({xs: xs_vv, ys: ys_vv})[ys_vv]
+    ys_logp = conditional_logp({xs: xs_vv, ys: ys_vv})[ys_vv]
 
     assert_no_rvs(ys_logp)
 
