@@ -47,13 +47,12 @@ from pytensor.tensor.random.basic import normal, uniform
 
 import pymc as pm
 
-from pymc.logprob.abstract import MeasurableVariable, get_measurable_outputs
-from pymc.logprob.basic import joint_logp, logp
+from pymc.logprob.abstract import MeasurableVariable
+from pymc.logprob.basic import logp, transformed_conditional_logp
 from pymc.logprob.utils import (
     ParameterValueError,
+    check_potential_measurability,
     dirac_delta,
-    ignore_logprob,
-    reconsider_logprob,
     rvs_to_value_vars,
     walk_model,
 )
@@ -198,65 +197,15 @@ def test_dirac_delta_logprob(dist_params, obs):
     scipy_logprob_tester(x, obs, dist_params, test_fn=scipy_logprob)
 
 
-def test_ignore_reconsider_logprob_basic():
-    x = pm.Normal.dist()
-    (measurable_x_out,) = get_measurable_outputs(x.owner.op, x.owner)
-    assert measurable_x_out is x.owner.outputs[1]
+def test_check_potential_measurability():
+    x1 = pt.random.normal()
+    x2 = pt.random.normal()
+    x3 = pt.scalar("x3")
+    y = pt.exp(x1 + x2 + x3)
 
-    new_x = ignore_logprob(x)
-    assert new_x is not x
-    assert isinstance(new_x.owner.op, pm.Normal)
-    assert type(new_x.owner.op).__name__ == "UnmeasurableNormalRV"
-    # Confirm that it does not have measurable output
-    assert get_measurable_outputs(new_x.owner.op, new_x.owner) == []
-
-    # Test that it will not clone a variable that is already unmeasurable
-    assert ignore_logprob(new_x) is new_x
-
-    orig_x = reconsider_logprob(new_x)
-    assert orig_x is not new_x
-    assert isinstance(orig_x.owner.op, pm.Normal)
-    assert type(orig_x.owner.op).__name__ == "NormalRV"
-    # Confirm that it has measurable outputs again
-    assert get_measurable_outputs(orig_x.owner.op, orig_x.owner) == [orig_x.owner.outputs[1]]
-
-    # Test that will not clone a variable that is already measurable
-    assert reconsider_logprob(x) is x
-    assert reconsider_logprob(orig_x) is orig_x
-
-
-def test_ignore_reconsider_logprob_model():
-    def custom_logp(value, x):
-        # custom_logp is just the logp of x at value
-        x = reconsider_logprob(x)
-        return joint_logp(
-            [x],
-            rvs_to_values={x: value},
-            rvs_to_transforms={},
-        )
-
-    with pm.Model():
-        x = pm.Normal.dist()
-        y = pm.CustomDist("y", x, logp=custom_logp)
-    with pytest.warns(
-        UserWarning,
-        match="Found a random variable that was neither among the observations "
-        "nor the conditioned variables",
-    ):
-        joint_logp(
-            [y],
-            rvs_to_values={y: y.type()},
-            rvs_to_transforms={},
-        )
-
-    # The above warning should go away with ignore_logprob.
-    with pm.Model():
-        x = ignore_logprob(pm.Normal.dist())
-        y = pm.CustomDist("y", x, logp=custom_logp)
-    with warnings.catch_warnings():
-        warnings.simplefilter("error")
-        assert joint_logp(
-            [y],
-            rvs_to_values={y: y.type()},
-            rvs_to_transforms={},
-        )
+    # In the first three cases, y is potentially measurable, because it has at least on unvalued RV input
+    assert check_potential_measurability([y], {})
+    assert check_potential_measurability([y], {x1})
+    assert check_potential_measurability([y], {x2})
+    # y is not potentially measurable because both RV inputs are valued
+    assert not check_potential_measurability([y], {x1, x2})
