@@ -41,9 +41,10 @@ import pymc as pm
 from pymc import Deterministic, Potential
 from pymc.blocking import DictToArrayBijection, RaveledVars
 from pymc.distributions import Normal, transforms
-from pymc.distributions.transforms import log
+from pymc.distributions.distribution import PartialObservedRV
+from pymc.distributions.transforms import log, simplex
 from pymc.exceptions import ImputationWarning, ShapeError, ShapeWarning
-from pymc.logprob.basic import joint_logp
+from pymc.logprob.basic import conditional_logp, transformed_conditional_logp
 from pymc.logprob.transforms import IntervalTransform
 from pymc.model import Point, ValueGradFunction, modelcontext
 from pymc.testing import SeededTest
@@ -356,13 +357,13 @@ class TestValueGradFunction(unittest.TestCase):
         gf = m.logp_dlogp_function()
         gf._extra_are_set = True
 
-        assert m["x2_missing"].type == gf._extra_vars_shared["x2_missing"].type
+        assert m["x2_unobserved"].type == gf._extra_vars_shared["x2_unobserved"].type
 
         # The dtype of the merged observed/missing deterministic should match the RV dtype
         assert m.deterministics[0].type.dtype == x2.type.dtype
 
         point = m.initial_point(random_seed=None).copy()
-        del point["x2_missing"]
+        del point["x2_unobserved"]
 
         res = [gf(DictToArrayBijection.map(Point(point, model=m))) for i in range(5)]
 
@@ -565,7 +566,7 @@ def test_make_obs_var():
     assert masked_output != fake_distribution
     assert not isinstance(masked_output, RandomVariable)
     # Ensure it has missing values
-    assert {"testing_inputs_missing"} == {v.name for v in fake_model.value_vars}
+    assert {"testing_inputs_unobserved"} == {v.name for v in fake_model.value_vars}
     assert {"testing_inputs", "testing_inputs_observed"} == {
         v.name for v in fake_model.observed_RVs
     }
@@ -1220,7 +1221,7 @@ class TestImputationMissingData:
             with pytest.warns(ImputationWarning):
                 _ = pm.Normal("y", x, 1, observed=missing_data)
 
-        assert "y_missing" in model.named_vars
+        assert "y_unobserved" in model.named_vars
 
         test_point = model.initial_point()
         assert not np.isnan(model.compile_logp()(test_point))
@@ -1237,7 +1238,7 @@ class TestImputationMissingData:
             with pytest.warns(ImputationWarning):
                 y = pm.Normal("y", x * predictors, 1, observed=data)
 
-        assert "y_missing" in model.named_vars
+        assert "y_unobserved" in model.named_vars
 
         test_point = model.initial_point()
         assert not np.isnan(model.compile_logp()(test_point))
@@ -1277,17 +1278,19 @@ class TestImputationMissingData:
             with pytest.warns(ImputationWarning):
                 theta2 = pm.Normal("theta2", mu=theta1, observed=obs2)
 
-            assert isinstance(model.rvs_to_transforms[model["theta1_missing"]], IntervalTransform)
+            assert isinstance(
+                model.rvs_to_transforms[model["theta1_unobserved"]], IntervalTransform
+            )
             assert model.rvs_to_transforms[model["theta1_observed"]] is None
 
             prior_trace = pm.sample_prior_predictive(random_seed=rng, return_inferencedata=False)
             assert set(prior_trace.keys()) == {
                 "theta1",
                 "theta1_observed",
-                "theta1_missing",
+                "theta1_unobserved",
                 "theta2",
                 "theta2_observed",
-                "theta2_missing",
+                "theta2_unobserved",
             }
 
             # Make sure the observed + missing combined deterministics have the
@@ -1302,14 +1305,16 @@ class TestImputationMissingData:
             # Make sure the missing parts of the combined deterministic matches the
             # sampled missing and observed variable values
             assert (
-                np.mean(prior_trace["theta1"][:, obs1.mask] - prior_trace["theta1_missing"]) == 0.0
+                np.mean(prior_trace["theta1"][:, obs1.mask] - prior_trace["theta1_unobserved"])
+                == 0.0
             )
             assert (
                 np.mean(prior_trace["theta1"][:, ~obs1.mask] - prior_trace["theta1_observed"])
                 == 0.0
             )
             assert (
-                np.mean(prior_trace["theta2"][:, obs2.mask] - prior_trace["theta2_missing"]) == 0.0
+                np.mean(prior_trace["theta2"][:, obs2.mask] - prior_trace["theta2_unobserved"])
+                == 0.0
             )
             assert (
                 np.mean(prior_trace["theta2"][:, ~obs2.mask] - prior_trace["theta2_observed"])
@@ -1325,18 +1330,22 @@ class TestImputationMissingData:
             )
             assert set(trace.varnames) == {
                 "theta1",
-                "theta1_missing",
-                "theta1_missing_interval__",
+                "theta1_unobserved",
+                "theta1_unobserved_interval__",
                 "theta2",
-                "theta2_missing",
+                "theta2_unobserved",
             }
 
             # Make sure that the missing values are newly generated samples and that
             # the observed and deterministic match
-            assert np.all(0 < trace["theta1_missing"].mean(0))
-            assert np.all(0 < trace["theta2_missing"].mean(0))
-            assert np.isclose(np.mean(trace["theta1"][:, obs1.mask] - trace["theta1_missing"]), 0)
-            assert np.isclose(np.mean(trace["theta2"][:, obs2.mask] - trace["theta2_missing"]), 0)
+            assert np.all(0 < trace["theta1_unobserved"].mean(0))
+            assert np.all(0 < trace["theta2_unobserved"].mean(0))
+            assert np.isclose(
+                np.mean(trace["theta1"][:, obs1.mask] - trace["theta1_unobserved"]), 0
+            )
+            assert np.isclose(
+                np.mean(trace["theta2"][:, obs2.mask] - trace["theta2_unobserved"]), 0
+            )
 
             # Make sure that the observed values are unchanged
             assert np.allclose(np.var(trace["theta1"][:, ~obs1.mask], 0), 0.0)
@@ -1377,7 +1386,7 @@ class TestImputationMissingData:
             with pytest.warns(ImputationWarning):
                 x = pm.Gamma("x", 1, 1, observed=[1, 1, 1, np.nan])
 
-        logp_val = m2.compile_logp()({"x_missing_log__": np.array([0])})
+        logp_val = m2.compile_logp()({"x_unobserved_log__": np.array([0])})
         assert logp_val == -4.0
 
     def test_missing_logp2(self):
@@ -1393,37 +1402,48 @@ class TestImputationMissingData:
                     "theta2", mu=theta1, observed=np.array([np.nan, np.nan, 2, np.nan, 4])
                 )
         m_missing_logp = m_missing.compile_logp()(
-            {"theta1_missing": [2, 4], "theta2_missing": [0, 1, 3]}
+            {"theta1_unobserved": [2, 4], "theta2_unobserved": [0, 1, 3]}
         )
 
         assert m_logp == m_missing_logp
 
-    def test_missing_multivariate(self):
-        """Test model with missing variables whose transform changes base shape still works"""
-
+    def test_missing_multivariate_separable(self):
         with pm.Model() as m_miss:
-            with pytest.raises(
-                NotImplementedError,
-                match="Automatic inputation is only supported for univariate RandomVariables",
-            ):
-                with pytest.warns(ImputationWarning):
-                    x = pm.Dirichlet(
-                        "x",
-                        a=[1, 2, 3],
-                        observed=np.array([[0.3, 0.3, 0.4], [np.nan, np.nan, np.nan]]),
-                    )
+            with pytest.warns(ImputationWarning):
+                x = pm.Dirichlet(
+                    "x",
+                    a=[1, 2, 3],
+                    observed=np.array([[0.3, 0.3, 0.4], [np.nan, np.nan, np.nan]]),
+                )
+        assert (m_miss["x_unobserved"].owner.op, pm.Dirichlet)
+        assert (m_miss["x_observed"].owner.op, pm.Dirichlet)
 
-        # TODO: Test can be used when local_subtensor_rv_lift supports multivariate distributions
-        # from pymc.distributions.transforms import simplex
-        #
-        # with pm.Model() as m_unobs:
-        #     x = pm.Dirichlet("x", a=[1, 2, 3])
-        #
-        # inp_vals = simplex.forward(np.array([0.3, 0.3, 0.4])).eval()
-        # assert np.isclose(
-        #     m_miss.compile_logp()({"x_missing_simplex__": inp_vals}),
-        #     m_unobs.compile_logp(jacobian=False)({"x_simplex__": inp_vals}) * 2,
-        # )
+        with pm.Model() as m_unobs:
+            x = pm.Dirichlet("x", a=[1, 2, 3], shape=(1, 3))
+
+        inp_vals = simplex.forward(np.array([[0.3, 0.3, 0.4]])).eval()
+        np.testing.assert_allclose(
+            m_miss.compile_logp(jacobian=False)({"x_unobserved_simplex__": inp_vals}),
+            m_unobs.compile_logp(jacobian=False)({"x_simplex__": inp_vals}) * 2,
+        )
+
+    def test_missing_multivariate_unseparable(self):
+        with pm.Model() as m_miss:
+            with pytest.warns(ImputationWarning):
+                x = pm.Dirichlet(
+                    "x",
+                    a=[1, 2, 3],
+                    observed=np.array([[0.3, 0.3, np.nan], [np.nan, np.nan, 0.4]]),
+                )
+
+        assert isinstance(m_miss["x_unobserved"].owner.op, PartialObservedRV)
+        assert isinstance(m_miss["x_observed"].owner.op, PartialObservedRV)
+
+        inp_values = np.array([0.3, 0.3, 0.4])
+        np.testing.assert_allclose(
+            m_miss.compile_logp()({"x_unobserved": [0.4, 0.3, 0.3]}),
+            st.dirichlet.logpdf(inp_values, [1, 2, 3]) * 2,
+        )
 
     def test_missing_vector_parameter(self):
         with pm.Model() as m:
@@ -1439,7 +1459,7 @@ class TestImputationMissingData:
         assert np.all(x_draws[:, 0] < 0)
         assert np.all(x_draws[:, 1] > 0)
         assert np.isclose(
-            m.compile_logp()({"x_missing": np.array([-10, 10, -10, 10])}),
+            m.compile_logp()({"x_unobserved": np.array([-10, 10, -10, 10])}),
             st.norm(scale=0.1).logpdf(0) * 6,
         )
 
@@ -1458,10 +1478,10 @@ class TestImputationMissingData:
         x_obs_rv = m["x_observed"]
         x_obs_vv = m.rvs_to_values[x_obs_rv]
 
-        x_unobs_rv = m["x_missing"]
+        x_unobs_rv = m["x_unobserved"]
         x_unobs_vv = m.rvs_to_values[x_unobs_rv]
 
-        logp = joint_logp(
+        logp = transformed_conditional_logp(
             [x_obs_rv, x_unobs_rv],
             rvs_to_values={x_obs_rv: x_obs_vv, x_unobs_rv: x_unobs_vv},
             rvs_to_transforms={},
@@ -1482,11 +1502,10 @@ class TestImputationMissingData:
                 x = pm.Normal("x", observed=data, dims=("observed",))
         assert model.named_vars_to_dims == {"x": ("observed",)}
 
-    def test_error_non_random_variable(self):
+    def test_symbolic_random_variable(self):
         data = np.array([np.nan] * 3 + [0] * 7)
         with pm.Model() as model:
-            msg = "x of type <class 'pymc.distributions.censored.CensoredRV'> is not supported"
-            with pytest.raises(NotImplementedError, match=msg):
+            with pytest.warns(ImputationWarning):
                 x = pm.Censored(
                     "x",
                     pm.Normal.dist(),
@@ -1494,6 +1513,10 @@ class TestImputationMissingData:
                     upper=10,
                     observed=data,
                 )
+        np.testing.assert_almost_equal(
+            model.compile_logp()({"x_unobserved": [0] * 3}),
+            st.norm.logcdf(0) * 10,
+        )
 
 
 class TestShared(SeededTest):
